Fieldcooling (#16)
* setting the entry type for field cooling

Probably there is a better way?

* Add samples on it's own

TODO: There is no checking if the sample already exists!
TODO: It should also probably be put into a normalizer

* forgot to add the utils file to git

* adds a lot of functionality for fieldcooling.

TODO: TESTS dont work, these need to be fixed!

* Removed the field cooling parser test

This is way too difficult to fix quickly.
I probably need to take the whole TextParsing out of the
field_cooling_parser, and then just test this TextParsing.

* adds description parsing for FC

* better plotly navigation

now the x-axis is not fixed anymore and zooming is better